--- marionette-v3.3.1-keyed/.babelrc
@@ -1,3 +1,0 @@
-{
-    "presets" : [ "babel-preset-es2015"]
-}
