+++ marionette-v3.4.0-keyed/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8"/>
-    <title>Marionette-v3.3.1</title>
+    <title>Marionette-v3.4.0</title>
     <link href="/css/currentStyle.css" rel="stylesheet"/>
 </head>
 <body>
@@ -11,7 +11,7 @@
         <div class="jumbotron">
             <div class="row">
                 <div class="col-md-6">
-                    <h1>Marionette-v3.3.1</h1>
+                    <h1>Marionette-v3.4.0</h1>
                 </div>
                 <div class="col-md-6">
                     <div class="row">
