+++ marionette-v3.4.0-keyed/package.json
@@ -1,7 +1,7 @@
 {
   "name": "js-framework-benchmark-marionette",
   "version": "1.0.0",
-  "description": "Marionette v3.1.0 demo",
+  "description": "Marionette v3.4.0",
   "main": "index.js",
   "scripts": {
     "build-dev": "webpack -w -d",
@@ -10,12 +10,13 @@
   "devDependencies": {
     "babel-core": "6.24.1",
     "babel-loader": "7.0.0",
+    "babel-plugin-handlebars-inline-precompile": "^2.1.1",
     "babel-preset-es2015": "6.24.1",
     "webpack": "2.5.1"
   },
   "dependencies": {
     "backbone": "1.3.3",
-    "backbone.marionette": "3.3.1",
+    "backbone.marionette": "3.4.0",
     "backbone.radio": "2.0.0",
     "jquery": "3.2.1",
     "underscore": "1.8.3"
