+++ marionette-v3.4.0-keyed/webpack.config.js
